sambar: camera: Modify thumbnail size logic
Previously the thumbnail size override logic was reverted because
it was causing failures in testJpegExif in cts2. However this
logic was apparently inadvertently fixing testJpegExif in the
original camera cts by modifying the thumbnail size selection
when the cts test requests a thumbnail larger than the picture
itself.

Change the thumbnail size logic to override the thumbnail size
only when the requested thumbnail is larger the picture itself.

Change-Id: I14ff846007ca5bdf0e8764566636674cc1d5e690
Signed-off-by: Tarun93 <tarun.kapadia@micromaxinfo.com>